Align latest k8s modules

+++ apis/config/v1alpha1/config_types.go
@@ -30,11 +30,12 @@ type ConfigSpec struct {
 	// will follow
 	Lifecycle *Lifecycle `json:"lifecycle,omitempty" protobuf:"bytes,1,opt,name=lifecycle"`
 	// Priority defines the priority of this config
-	Priority int64 `json:"priority,omitempty" protobuf:"varint,2,opt,name=priority"`
+	Priority int32 `json:"priority,omitempty" protobuf:"varint,2,opt,name=priority"`
 	// Revertive defines if this CR is enabled for revertive or non revertve operation
 	Revertive *bool `json:"revertive,omitempty" protobuf:"varint,3,opt,name=revertive"`
 	// Config defines the configuration to be applied to a target device
-	//+kubebuilder:pruning:PreserveUnknownFields
+	// +kubebuilder:pruning:PreserveUnknownFields
+	// +listType=atomic
 	Config []ConfigBlob `json:"config" protobuf:"bytes,4,rep,name=config"`
 }
